pyproject.toml
[tool.poetry]
name = "understory"
version = "0.0.63"
description = "The tools that power the canopy"
authors = ["Angelo Gladding <angelo@lahacker.net>"]
license = "AGPL-3.0-or-later"
readme = "README.md"

[tool.poetry.urls]
Homepage = "https://understory.stream"
Issues = "https://github.com/canopy/understory/issues"

[tool.poetry.scripts]
loveliness = "understory.loveliness:main"
# indieauth = "understory.indieweb.indieauth:main"
# micropub = "understory.indieweb.micropub:main"
# microsub = "understory.indieweb.microsub:main"
# followers = "understory.indieweb.followers:main"
# mentions = "understory.indieweb.webmention:main"
pkg = "understory.pkg.__main__:main"
web = "understory.web.__main__:main"

[tool.poetry.plugins."web.apps"]
content = "understory.indieweb:content"
# db = "understory.indieweb:db"
# cache = "understory.indieweb:cache"
# favicon = "understory.indieweb:favicon"
# static = "understory.indieweb:static"
# templates = "understory.indieweb:templates"
"indieauth-client" = "understory.indieweb.indieauth:client"
"indieauth-server" = "understory.indieweb.indieauth:server"
micropub = "understory.indieweb.micropub:server"
microsub = "understory.indieweb.microsub:server"
webmention = "understory.indieweb.webmention:receiver"
websub = "understory.indieweb.websub:hub"

[tool.poetry.dependencies]
python = "^3.9,<3.10"
argcomplete = "^1.12.3"
radon = "^5.0.1"
pendulum = "^2.1.2"
sh = "^1.14.2"
semver = "^2.13.0"
lxml = "^4.6.3"
Pygments = "^2.9.0"
emoji = "^1.2.0"
redis = "^3.5.3"
gunicorn = "^20.1.0"
hstspreload = "^2021.7.5"
requests = "^2.25.1"
mf2py = "^1.1.2"
mf2util = "^0.5.1"
networkx = "^2.6.1"
pyscreenshot = "^3.0"
Pillow = "^8.3.1"
PyVirtualDisplay = "^2.2"
selenium = "^3.141.0"
gevent = "^21.1.2"
jsonpatch = "^1.32"
pycryptodome = "^3.10.1"
Unidecode = "^1.2.0"
watchdog = "^2.1.3"
httpagentparser = "^1.9.1"
mimeparse = "^0.1.3"
acme-tiny = "^4.1.0"
scrypt = "^0.8.18"
vobject = "^0.9.6"
cssselect = "^1.1.0"
dnspython = "^2.1.0"
feedparser = "^6.0.8"
regex = "^2021.7.6"
rich = "^10.7.0"

[tool.poetry.dev-dependencies]
ipython = "^7.25.0"
pytest = "^6.2.4"
pytest-cov = "^2.12.1"
webdriver-manager = "^3.4.2"
black = "^21.6b0"
pytest-gevent = "^1.1.0"
isort = "^5.9.2"
flake8 = "^3.9.2"
flake8-docstrings = "^1.6.0"
pep8-naming = "^0.12.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
